WeatherAPI

+++ WeatherApiApplication/MyClassLibrary/Clouds.cs
@@ -2,6 +2,6 @@
 {
     public class Clouds
     {
-        public int all { get; set; }
+        public double all { get; set; }
     }
 }
